+++ crates/e2e/src/subxt_client.rs
@@ -46,10 +46,7 @@ use crate::{
         ContractResult,
         UploadResult,
     },
-    error::{
-        DryRunError,
-        DryRunRevert,
-    },
+    error::DryRunError,
     events,
     ContractsBackend,
     E2EBackend,
@@ -162,6 +159,7 @@ where
         storage_deposit_limit: E::Balance,
     ) -> Result<BareInstantiationResult<ExtrinsicEvents<C>>, Error> {
         let salt = salt();
+        // todo remove assert once salt() returns no more option
         assert!(salt.is_some());
         let (events, trace) = self
             .api
@@ -549,6 +547,7 @@ where
             .contract_result_to_result(result)
             .map_err(Error::InstantiateDryRun)?;
 
+        /*
         if let Ok(res) = result.result.clone() {
             if res.result.did_revert() {
                 return Err(Self::Error::InstantiateDryRunReverted(DryRunRevert {
@@ -556,6 +555,7 @@ where
                 }));
             }
         }
+         */
 
         Ok(result.into())
     }
@@ -846,8 +846,6 @@ impl<E: Environment, V, C: subxt::Config> CallResult<E, V, ExtrinsicEvents<C>> {
                 let topics = decoded_event.topics.clone();
                 let event_with_topics = EventWithTopics {
                     event: decoded_event,
-                    //topics: event.topics().iter().cloned().map(Into::into).collect(),
-                    //topics: topics.iter().map(|v| H256::from_slice(&v[..])).collect(),
                     topics,
                 };
                 events_with_topics.push(event_with_topics);